fixed the errors relating to contentjsx, counterjsx,galleryjsx, and mainjsx
