--- src/components/Content.jsx
@@ -1,43 +1,0 @@
-function Gallery() {
-//   let deer = {
-//       title: "Deer",
-//       imageUrl: "../images/deer.jpg",
-//       description: "This is bambi",
-//   };
-
-//   let antelope = {
-//       title: "Antelope",
-//       imageUrl: "../images/caribou.jpg",
-//       description: "This is an antelope.",
-//   };
-
-//   let bighornSheep = {
-//       title: "Bighorn Sheep",
-//       imageUrl: "../images/bighorn.jpg",
-//       description: "This is a bighorn sheep.",
-//   };
-
-  return (
-      <main>
-          {/* <section className="deck">
-              <div className="card">
-                  <h2>{deer.title}</h2>
-                  <img src={deer.image} alt={deer.title} />
-                  <p>{deer.description}</p>
-              </div>
-              <div className="card">
-                  <h2>{antelope.title}</h2>
-                  <img src={antelope.imageUrl} alt={antelope.title} />
-                  <p>{antelope.description}</p>
-              </div>
-              <div className="card">
-                  <h2>{bighornSheep.title}</h2>
-                  <img src={bighornSheep.imageUrl} alt={bighornSheep.title} />
-                  <p>{bighornSheep.description}</p>
-              </div>
-          </section> */}
-      </main>
-  );
-}
-
-export default Gallery;
